chore: refactor of repositories
