+++ frontnuxt/lib/config.ts
@@ -1,0 +1,1 @@
+export const BACKEND_HOST = process.env.BACKEND_HOST || 'http://localhost:8080';
